fix a variety of masked texture bugs

+++ old_c/r_data.c
@@ -2771,7 +2771,7 @@ segment_t __far R_GetMaskedColumnSegment (int16_t tex, int16_t col) {
 		// "modulo from negative" by patch width.
 
 		if (col < 0){
-			uint16_t patchwidth = patchwidths_7000[lump-firstpatch];
+			uint8_t patchwidth = patchwidths_7000[lump-firstpatch];
 			if (patchwidth == 0){
 				patchwidth = 0x100;
 			}
@@ -2802,7 +2802,7 @@ segment_t __far R_GetMaskedColumnSegment (int16_t tex, int16_t col) {
 
 			return maskedcachedsegment + ofs;
 		}
-	} else {
+	} else { 
 		uint8_t collength = texturecollength[tex];
 
 		// todo in the asm make default branch to use cache
